requirements.txt
arrow==0.8.0
Django==1.9.8
djangorestframework==3.4.0
psycopg2==2.6.2
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
six==1.10.0
stravalib==0.6.0
units==0.6
